Docker compose (#19)
* dockerize: build a fully functional tracking_services within a docker image (#1)

* created two dockerfiles

Dockerfile.local builds the current repository in whatever state it is in. It copies the files from the repo into the container meaning that the files do not even need to be committed to be tested.

Dockerfile.release builds from the source code of a release. It requires the argument that defines the version to be downloaded.

* updated pom.xml

To maven build with JDK 11 and jaxb had to add a jaxb dependency to the pom and define UTF-8 as the encoding.

* Update README.md

* Update README.md

* update fork (#2)

* docker compose starts trakcing service and mysql, not tested

* use docker compose for docker deployment, make db connection work

* ad instruction for release build in docker

* docker compose starts trakcing service and mysql, not tested

* use docker compose for docker deployment, make db connection work

* ad instruction for release build in docker

Co-authored-by: al-niessner <1130658+al-niessner@users.noreply.github.com>
Co-authored-by: thomas loubrieu <thomas.loubrieu@jpl.nasa.gov>